chore: 프로젝트 초기 환경설정 및 Swift Lint 적용
