+++ Diary/Application/SceneDelegate.swift
@@ -1,0 +1,22 @@
+//
+//  Diary - SceneDelegate.swift
+//  Created by Minseong, Lingo
+// 
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+  var window: UIWindow?
+
+  func scene(
+    _ scene: UIScene,
+    willConnectTo session: UISceneSession,
+    options connectionOptions: UIScene.ConnectionOptions
+  ) {
+    guard let _ = (scene as? UIWindowScene) else { return }
+  }
+
+  func sceneDidEnterBackground(_ scene: UIScene) {
+    (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+  }
+}
