+++ Diary/Controller/ViewController.swift
@@ -1,0 +1,12 @@
+//
+//  Diary - ViewController.swift
+//  Created by Minseong, Lingo
+//
+
+import UIKit
+
+final class ViewController: UIViewController {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
