+++ Diary/Resource/Assets.xcassets/AccentColor.colorset/Contents.json
