+++ Diary/Resource/Assets.xcassets/AppIcon.appiconset/Contents.json
