+++ Diary/Resource/Assets.xcassets/Contents.json
