+++ Diary/Resource/Base.lproj/LaunchScreen.storyboard
