+++ Diary/Resource/Diary.xcdatamodeld/Diary.xcdatamodel/contents
